// Client/src/app/references/dataitems/attorneys.ts
export interface AttorneyItem {
    'attorneyId': number;
    'attorneyName': string;
    'address2': string;
    'address1': string;
    'city': string;
    'postalCode': string;
    'stateId': number;
    'stateName': string;
    'extension': string;
    'faxNumber': string;
    'phoneNumber': string;
}